Added TypeScripts + iOS Android fixes/refactoring Refactoring entailed renaming some variables and moving event callbacks from an options object to method arguments. Fixes entailed converting file-stat-related fields to doubles as to avoid MaxInteger bugs in reading file sizes.

+++ android/src/main/java/com/rnfs/DownloadParams.java
@@ -12,11 +12,11 @@ public class DownloadParams {
   }
 
   public interface OnDownloadBegin {
-    void onDownloadBegin(int statusCode, int contentLength, Map<String, String> headers);
+    void onDownloadBegin(int statusCode, double contentLength, Map<String, String> headers);
   }
 
   public interface OnDownloadProgress {
-    void onDownloadProgress(int contentLength, int bytesWritten);
+    void onDownloadProgress(double contentLength, double bytesWritten);
   }
 
   public URL src;
